Uses oslo.service for ceilometer-polling on Windows
Recently, ceilometer switched from using oslo.service to
cotyledon, which is not supported on Windows, as it uses
Linux-specific attributes and functions, causing the service
to crash.

This patch ensures that oslo.service is used on Windows for
ceilometer-polling, while maintaining the current usage of
cotyledon on any other platform.

This patch will have to reverted once cotyledon works properly
on Windows.

Change-Id: I504289d9470fafb261bed587768b0527d4b3ecca
Partial-Bug: #1625602